etl/requirements.txt
# ETL环境依赖
mysql-connector-python==8.0.33
